backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
python-dotenv
joblib
requests
pydantic
openai
paho-mqtt
firebase-admin
